fix\EV-739\Removing Componetization from HomePage
