--- lib/screens/home_page/widgets/header.dart
@@ -1,94 +1,0 @@
-import 'dart:ui';
-
-import 'package:crypto/data/crypto_list_repository.dart';
-import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:intl/intl.dart';
-
-import '../../../shared/providers/providers.dart';
-import '../../../utils/constants.dart';
-
-class Header extends ConsumerStatefulWidget {
-  const Header({Key? key}) : super(key: key);
-
-  @override
-  ConsumerState<Header> createState() => _HeaderState();
-}
-
-String walletBalance() {
-  final formater = NumberFormat("#,##0.00", "pt");
-  CryptoListRepository repository = CryptoListRepository();
-  double wallet = 0;
-  for (var crypto in repository.cryptoListRepository) {
-    wallet += crypto.userBalance;
-  }
-  return formater.format(wallet);
-}
-
-class _HeaderState extends ConsumerState<Header> {
-  @override
-  Widget build(BuildContext context) {
-    final bool visibility = ref.watch(visibilityProvider);
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: <Widget>[
-        Padding(
-          padding: kDefaultPadding,
-          child: Column(
-            children: [
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  const Text(
-                    'Cripto',
-                    style: kTitleTextStyle,
-                  ),
-                  IconButton(
-                    onPressed: () {
-                      ref.read(visibilityProvider.notifier).state = !visibility;
-                      setState(() {});
-                    },
-                    icon: Icon(
-                      visibility
-                          ? Icons.visibility_rounded
-                          : Icons.visibility_off_rounded,
-                      size: 26,
-                    ),
-                  )
-                ],
-              ),
-              const SizedBox(height: 10),
-              Row(
-                children: [
-                  const Text(
-                    "R\$ ",
-                    style: kWalletBalanceTextStyle,
-                  ),
-                  ImageFiltered(
-                    imageFilter: visibility
-                        ? ImageFilter.blur()
-                        : ImageFilter.blur(sigmaX: 8, sigmaY: 8),
-                    child: Text(
-                      walletBalance(),
-                      style: kWalletBalanceTextStyle,
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 10),
-              Row(
-                children: const [
-                  Text(
-                    "Valor total de moedas",
-                    style: kSubtitleTextStyle,
-                  ),
-                ],
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-}
